Added sidebar for coin transfer
