+++ common/config/rush/repo-state.json
@@ -1,6 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-
-  "pnpmShrinkwrapHash": "aa6137db4c6b3e3e060557e38700abb4bc50a064",
+  "pnpmShrinkwrapHash": "a957f87bcd732f3bd5354a39ca3f8c4cd267c188",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
